+++ go/tasks/plugins/presto/config/config_flags_test.go
@@ -143,50 +143,6 @@ func TestConfig_SetFlags(t *testing.T) {
 			}
 		})
 	})
-	t.Run("Test_awsS3ShardFormatter", func(t *testing.T) {
-		t.Run("DefaultValue", func(t *testing.T) {
-			// Test that default value is set properly
-			if vString, err := cmdFlags.GetString("awsS3ShardFormatter"); err == nil {
-				assert.Equal(t, string(defaultConfig.AwsS3ShardFormatter), vString)
-			} else {
-				assert.FailNow(t, err.Error())
-			}
-		})
-
-		t.Run("Override", func(t *testing.T) {
-			testValue := "1"
-
-			cmdFlags.Set("awsS3ShardFormatter", testValue)
-			if vString, err := cmdFlags.GetString("awsS3ShardFormatter"); err == nil {
-				testDecodeJson_Config(t, fmt.Sprintf("%v", vString), &actual.AwsS3ShardFormatter)
-
-			} else {
-				assert.FailNow(t, err.Error())
-			}
-		})
-	})
-	t.Run("Test_awsS3ShardCount", func(t *testing.T) {
-		t.Run("DefaultValue", func(t *testing.T) {
-			// Test that default value is set properly
-			if vInt, err := cmdFlags.GetInt("awsS3ShardCount"); err == nil {
-				assert.Equal(t, int(defaultConfig.AwsS3ShardCount), vInt)
-			} else {
-				assert.FailNow(t, err.Error())
-			}
-		})
-
-		t.Run("Override", func(t *testing.T) {
-			testValue := "1"
-
-			cmdFlags.Set("awsS3ShardCount", testValue)
-			if vInt, err := cmdFlags.GetInt("awsS3ShardCount"); err == nil {
-				testDecodeJson_Config(t, fmt.Sprintf("%v", vInt), &actual.AwsS3ShardCount)
-
-			} else {
-				assert.FailNow(t, err.Error())
-			}
-		})
-	})
 	t.Run("Test_rateLimiter.name", func(t *testing.T) {
 		t.Run("DefaultValue", func(t *testing.T) {
 			// Test that default value is set properly
